% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Calculate Distance from Latitude and Longitude}
\usage{
distance(lat1, long1, lat2, long2)
}
\arguments{
\item{lat1}{Latitude of the first point in degrees.}

\item{long1}{Longitude of the first point in degrees.}

\item{lat2}{Latitude of the second point in degrees.}

\item{long2}{Longitude of the second point in degrees.}
}
\value{
The distance in kilometers between the two points.
}
\description{
This function calculates the great-circle distance between two points on Earth
given their latitude and longitude coordinates.
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP} and modified for use in R-Instat.
The modifications include changes to data input and output processes while carefully maintaining consistent calculations
processes as in the original files.This function calculates the great-circle distance using the Haversine formula.
It takes latitude and longitude values in degrees and returns the distance in kilometers.
}
\examples{
# Calculate the distance between two points
# distance(40.7128, -74.0060, 34.0522, -118.2437)

}
\references{
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
}
