% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{Exponential}
\alias{Exponential}
\title{Exponential Distribution Calculator}
\usage{
Exponential(x, n, r, s)
}
\arguments{
\item{x}{A numeric vector or scalar containing input values.}

\item{n}{The minimum value of the distribution.}

\item{r}{The rate parameter, controlling the rate of decay.}

\item{s}{The scale factor, affecting the overall shape.}
}
\value{
A numeric vector of the same length as \code{x}, representing the probability values of the exponential distribution.
}
\description{
Calculate the probability values of an exponential distribution with user-defined parameters.
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP} and modified for use in R-Instat.
The modifications include changes to data input and output processes while preserving the core calculation methods.
The exponential distribution is commonly used to model the time between events in a Poisson process.
The parameters \code{n}, \code{r}, and \code{s} determine the shape and characteristics of the distribution.
}
\examples{
# Generate an exponential distribution with minimum value 0, rate parameter 1, and scale 2
# x <- seq(0, 5, by = 0.1)
# y <- Exponential(x, 0, 1, 2)
# plot(x, y, type = "l", main = "Exponential Distribution")

}
\references{
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
}
