% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian.R
\name{Gaussian}
\alias{Gaussian}
\title{Gaussian Function}
\usage{
Gaussian(x, n, r, s)
}
\arguments{
\item{x}{A numeric vector or scalar containing input values.}

\item{n}{The center of the Gaussian curve.}

\item{r}{The standard deviation (width) of the Gaussian curve.}

\item{s}{The scale (height) of the Gaussian curve.}
}
\value{
A numeric vector of the same length as x, representing the Gaussian distribution.
}
\description{
This function calculates a Gaussian (normal) distribution with parameters n, r, and s,
for a given input vector x.
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP} and modified for use in R-Instat.
The modifications include changes to data input and output processes while preserving the core calculation methods.
The Gaussian function is used to model a normal distribution with parameters n, r, and s.
It describes the probability distribution of a continuous random variable.
}
\examples{
# Generate a Gaussian curve with center at 0, standard deviation 1, and scale 2
# x <- seq(-3, 3, by = 0.1)
# y <- Gaussian(x, 0, 1, 2)
# plot(x, y, type = "l", main = "Gaussian Distribution")

}
\references{
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
}
