% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_indices.R
\name{p2_indices}
\alias{p2_indices}
\title{Calculate Climatological Monthly and Annual Indices}
\usage{
p2_indices(
  df,
  station,
  date,
  precip,
  tmax,
  tmin,
  qct = 0,
  qcpr = 0,
  nybr = 1981,
  nyer = 2010,
  station_df,
  name,
  lat,
  lon
)
}
\arguments{
\item{df}{A data.frame containing daily climate data.}

\item{station}{The name of the station column in the data frame.}

\item{date}{The name of the date column in the data frame.}

\item{precip}{The name of the precipitation/rain column in the data frame.}

\item{tmax}{The name of the maximum temperature column in the data frame.}

\item{tmin}{The name of the minimum temperature column in the data frame.}

\item{qct}{Quality Control level for temperature data (0, 1, or 2).}

\item{qcpr}{Quality Control level for precipitation data (0, 1, or 2).}

\item{nybr}{The start year for the climatological period.}

\item{nyer}{The end year for the climatological period.}

\item{station_df}{A data.frame containing station metadata.}

\item{name}{The name of the station column in station data frame (to link with df).}

\item{lat}{The name of the latitude column in station data frame.}

\item{lon}{The name of the longitude column in station data frame.}
}
\value{
A data frame containing calculated monthly and annual climate indices.
}
\description{
This function calculates various climatological monthly and annual climate indices
using daily climate data.The indices include mean temperature, temperature anomalies, precipitation totals,
precipitation anomalies, standardized precipitation index (SPI), and various temperature percentiles.
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP} and modified for use in R-Instat.
The modifications include changes to data input and output processes while preserving the core calculation methods.
The function calculates various climate indices for each station in the given data frame.It requires daily climate data,
including maximum temperature (tmax), minimum temperature (tmin),and precipitation (precip). The function calculates monthly
and annual indices, including mean temperature,temperature anomalies, precipitation totals, precipitation anomalies,
standardized precipitation index (SPI),and various temperature percentiles. The climatology is defined by
the \code{nybr} (start year) and \code{nyer} (end year) parameters.
}
\examples{
# Example usage:
# indices_result <- p2_indices(df = climate_data, station = "StationID", date = "Date",
# precip = "Precipitation", tmax = "MaxTemperature",
# tmin = "MinTemperature", qcpr = 2, nybr = 1981, nyer = 2010,
# station_df = station_metadata, name = "StationName",
# lat = "Latitude", lon = "Longitude")

}
\references{
reference
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
}
