% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P3_variogram.R
\name{p3_variogram}
\alias{p3_variogram}
\title{Calculate Variograms for Climate Indices}
\usage{
p3_variogram(
  a2,
  station,
  year,
  month,
  nyb = 1981,
  nye = 2010,
  ne = 1:8,
  station_df,
  name,
  lat,
  lon,
  ncmp_folder
)
}
\arguments{
\item{a2}{A data frame containing climate indices data.}

\item{station}{Name of the station column in 'a2'.}

\item{year}{Name of the year column in 'a2'.}

\item{month}{Name of the month column in 'a2'.}

\item{nyb}{Start year for the variogram period.}

\item{nye}{End year for the variogram period.}

\item{ne}{(Optional) Specifies which climate indices to compute variograms for (1 to 8).}

\item{station_df}{A data frame with station metadata.}

\item{name}{Name of the station column in 'station_df' (to link with 'a2' using 'station').}

\item{lat}{Name of the latitude column in 'station_df'.}

\item{lon}{Name of the longitude column in 'station_df'.}

\item{ncmp_folder}{Path to the folder containing NCMP data.}
}
\value{
A data frame containing variogram information for each climate index.
}
\description{
This function computes variograms for climate indices based on station data.
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP} and modified for use in R-Instat.
The modifications include changes to data input and output processes while preserving the core calculation methods.
The function calculates variograms for climate indices, which are used in geostatistical analysis.
Variograms quantify spatial variability and help in predicting values at unobserved locations.
The function computes variograms for different climate indices specified by 'ne' and returns
information about the variogram models fitted to the data.
}
\examples{
# Example usage of p3_variogram:
# Assuming 'a2', 'station', 'year', 'month', 'station_df', 'name', 'lat', 
#            'lon', and 'ncmp_folder' are defined.
# result <- p3_variogram(a2, station, year, month, nyb = 1981, nye = 2010, 
#  ne = 1:8,station_df, name, lat, lon, ncmp_folder)

}
\references{
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
}
