% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p4_region_average_function.R
\name{p4_region_average}
\alias{p4_region_average}
\title{Calculate Regional Average of Climate Indices}
\usage{
p4_region_average(
  a2,
  station,
  year,
  month,
  a3,
  ne = 1:8,
  station_df,
  name,
  lat,
  lon,
  nyb = 1950,
  nye,
  uncode,
  label,
  res,
  igrid = FALSE
)
}
\arguments{
\item{a2}{Data frame containing climate indices from p2_indices().}

\item{station}{Name of the station column in a2.}

\item{year}{Name of the year column in a2.}

\item{month}{Name of the month column in a2.}

\item{a3}{Data frame containing variogram output from p3_variogram().}

\item{ne}{(Optional) Vector specifying which indices to compute regional averages for, ranging from 1 to 8.}

\item{station_df}{Data frame with station metadata.}

\item{name}{Name of the station column in station_df (to link with a2).}

\item{lat}{Name of the latitude column in station_df.}

\item{lon}{Name of the longitude column in station_df.}

\item{nyb}{Start year.}

\item{nye}{End year.}

\item{uncode}{UN country code for the region as defined in the UN country code list (0 for other regions).}

\item{label}{(Optional) An alternative name for the selected country or region.}

\item{res}{Grid spacing in degrees for the interpolation. Should be one of 0.1, 0.25, 0.5, 1, or 2.}

\item{igrid}{If TRUE, output grid square data.}
}
\value{
A list containing regional average data frames and optionally grid square data frames if igrid is set to TRUE.
}
\description{
This function calculates the regional average of 8 different climate indices.
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
and modified for use in R-Instat.The modifications include changes to data input and output
processes while preserving the core calculation methods.This function calculates the regional
average of 8 different indices.
}
\examples{
# Example usage:
# data <- p4_region_average(a2_data, "station_col", "year_col", "month_col",
#                           a3_data, ne = 1:8, station_df = station_metadata, 
#                           name = "station_name", lat = "latitude", lon = "longitude", 
#                           nyb = 1950, nye = 2020, uncode = 840, res = 1, igrid = FALSE)

}
\references{
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
}
