% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p5_trends_graphs_function.R
\name{p5_trends_graphs}
\alias{p5_trends_graphs}
\title{Calculating Trends Graphs}
\usage{
p5_trends_graphs(
  a2,
  station,
  a2_year,
  month,
  a4,
  a4_year,
  a4_month,
  ne = 1:8,
  nyba,
  nyea,
  nyb,
  nye,
  ncmp_folder,
  station_df,
  name,
  lat,
  lon
)
}
\arguments{
\item{a2}{data.frame containing 21 indices from p2_indices()}

\item{station}{name of station column in a2}

\item{a2_year}{name of year column in a2}

\item{month}{name of month column in a2}

\item{a4}{data.frame containing main output data.frame() from p4_region_average()}

\item{a4_year}{name of year column in a4}

\item{a4_month}{name of month column in a4}

\item{ne}{(optional) which indices to compute region average for, 1 to 8.}

\item{nyba}{(optional) start year for region average (if missing will use beginning year from a4)}

\item{nyea}{(optional) end year for region average (if missing will use end year from a4)}

\item{nyb}{(optional) start year for station trend (if missing will use nyba)}

\item{nye}{(optional) end year for station trend (if missing will use nyea)}

\item{ncmp_folder}{directory to output graphs and maps. ncmp_folder/A5_Trends_Graphs will be created for output.}

\item{station_df}{data.frame will station metadata}

\item{name}{name of station column in station_df (to link with a2)}

\item{lat}{name of latitude column in station_df}

\item{lon}{name of longitude column in station_df}
}
\value{
A list containing data frames with trends and significance values for both stations and regions.
}
\description{
This function reads the indices calculated previously and calculates the trend for
the period specified by the user.It produces a map of each of these trends by station.
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP} and modified for use in R-Instat.
The modifications include changes to data input and output processes while preserving the core calculation methods.
This function reads the indices calculated previously and calculates the trend for the period specified by the user.
It produces a map of each of these trends by station.
}
\examples{
# Example usage of 'p5_trends_graphs':
# p5_trends_graphs(a2, station = "StationID", a2_year = "Year", month = "Month",
#                  a4, a4_year = "Year", a4_month = "Month",ne = 1:8, nyba = 1950,
#                  nyea = 2020, nyb = 1960, nye = 2020, ncmp_folder = "output_folder",
#                station_df, name = "StationName", lat = "Latitude", lon = "Longitude")

}
\references{
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
}
