% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p6_count_records.R
\name{p6_count_records}
\alias{p6_count_records}
\title{Counting the Number of Records for Multiple Indices}
\usage{
p6_count_records(
  a2,
  station,
  a2_year,
  month,
  a4,
  nyb,
  nye,
  station_df,
  name,
  lat,
  lon
)
}
\arguments{
\item{a2}{A data frame containing climate indices data.}

\item{station}{The name of the station column in the data frame.}

\item{a2_year}{The name of the year column in the data frame.}

\item{month}{The name of the month column in the data frame.}

\item{a4}{A data frame containing region-average data.}

\item{nyb}{(optional) The start year for the count period (default is the beginning year from a4).}

\item{nye}{(optional) The end year for the count period (default is the end year from a4).}

\item{station_df}{A data frame with station metadata.}

\item{name}{The name of the station column in the station metadata (to link with a2).}

\item{lat}{The name of the latitude column in the station metadata.}

\item{lon}{The name of the longitude column in the station metadata.}
}
\value{
A data frame containing counts of records for each index and each year.
}
\description{
This function counts the number of records for five different indices in a given dataset.
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP} and modified for use in R-Instat.
The modifications include changes to data input and output processes while preserving the core calculation methods.
This function counts the number of records for five climate indices, including extreme cold and warm temperatures
and extreme precipitation events. It calculates the counts based on the input data frames, station metadata, and specified time periods.
The function iterates through stations, assesses the availability of data, and computes the record counts for
the specified indices and years. It provides a summary of the data completeness for each station and index.
}
\examples{
# Example usage:
# counts <- p6_count_records(a2, "Station", "Year", "Month", a4)

}
\references{
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
}
