% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p7_summary.R
\name{p7_summary}
\alias{p7_summary}
\title{Assembling NCMPs for WMO Submission}
\usage{
p7_summary(a2, a4, a6, nyb, nye, ncmp_folder)
}
\arguments{
\item{a2}{Data frame containing station information and indices.}

\item{a4}{Data frame containing regional average information.}

\item{a6}{Data frame containing count records.}

\item{nyb}{Start year for the summary file period.}

\item{nye}{End year for the summary file period.}

\item{ncmp_folder}{The folder where NCMPs will be saved.}
}
\value{
A data frame containing assembled NCMPs ready for submission.
}
\description{
This function assembles the National Climate Monitoring Products (NCMPs) to prepare them for submission to the World Meteorological Organization (WMO).
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP} and modified for use in R-Instat.
The modifications include changes to data input and output processes while preserving the core calculation methods.
This function assembles the National Climate Monitoring Products (NCMPs) to prepare them for submission to the World Meteorological Organization (WMO).
}
\examples{
# Example usage:
# ncmp_data <- p7_summary(a2_data, a4_data, a6_data, 1980, 2020, "NCMP_Folder")

}
\references{
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
}
