% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_days_op_threshold_mod.R
\name{percent.days.op.threshold.mod}
\alias{percent.days.op.threshold.mod}
\title{Calculate Percent Days Diagnostics in the Presence of Limited Data}
\usage{
percent.days.op.threshold.mod(ci, freq, type)
}
\arguments{
\item{ci}{A climdexInput object containing climate data.}

\item{freq}{The frequency of the analysis, either "monthly" or "annual".}

\item{type}{The type of percentile analysis to perform, one of "TN10p", "TN90p", "TX10p", or "TX90p".}
}
\value{
A numeric vector containing the calculated percent days diagnostics.
}
\description{
This function calculates percent days diagnostics when there is an insufficient number of years available for analysis.
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP/}{ET-NCMP/NCMP project} and modified for use in R-Instat.
It specifically corrects the calculation of percent days diagnostics in cases where there are too few years of data.
Previous versions generated warnings and unrealistic negative bias in the estimated values.
The modifications include changes to data input and output processes while preserving the core calculation methods.
This function directly calculates percent days diagnostics and circumvents most climdex.pcic wrapper functions for ease of use.
}
\examples{
# Example usage:
# data <- your_climdexInput_object
# result <- percent.days.op.threshold.mod(data, freq = "monthly", type = "TN10p")

# The 'result' vector will contain the calculated percent days diagnostics.

}
\references{
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP/}{ET-NCMP/NCMP project}
}
