% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spherical.R
\name{Spherical}
\alias{Spherical}
\title{Spherical Interpolation Function}
\usage{
Spherical(x, n, r, s)
}
\arguments{
\item{x}{A numeric vector or scalar containing input values.}

\item{n}{The starting value on the unit sphere.}

\item{r}{The rotation angle in radians.}

\item{s}{The ending value on the unit sphere.}
}
\value{
A numeric vector of the same length as \code{x}, representing the slerp interpolation between \code{n} and \code{s}.
}
\description{
Spherical interpolation (slerp) is a method for smoothly interpolating between two values on a unit sphere.
This function calculates slerp for a given input vector \code{x} with user-defined parameters.
}
\details{
This function is adapted from the \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP} and modified for use in R-Instat.
The modifications include changes to data input and output processes while preserving the core calculation methods.
Slerp is often used in computer graphics and animation to smoothly interpolate between two rotations or orientations.
This function calculates slerp for a given input vector \code{x} based on the parameters \code{n} (start value), \code{r} (rotation angle),
and \code{s} (end value). It ensures that the interpolation stays on the unit sphere.
}
\examples{
# Perform spherical interpolation between (0, 0, 1) and (1, 0, 0) with rotation of pi/2
# x <- seq(0, 1, by = 0.1)
# n <- c(0, 0, 1)
# r <- pi / 2
# s <- c(1, 0, 0)
# result <- Spherical(x, n, r, s)
# plot(result, type = "l", main = "Spherical Interpolation")

}
\references{
For the original source code and more information, please refer to: \href{https://github.com/ET-NCMP/NCMP}{ET-NCMP/NCMP}
}
